requirements.txt
Flask==3.0.3
requests==2.31.0
APScheduler==3.10.4
Pillow==10.4.0
gunicorn==22.0.0
